+++ src/main/java/spring/beautiq/domain/makeup/dto/RecommendRequestDto.java
@@ -1,4 +1,4 @@
-package spring.beautiq.makeUp.dto;
+package spring.beautiq.domain.makeup.dto;
 
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
