+++ src/main/java/spring/beautiq/domain/makeup/entity/MakeUp.java
@@ -1,4 +1,4 @@
-package spring.beautiq.makeUp.entity;
+package spring.beautiq.domain.makeup.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
