+++ src/main/java/spring/beautiq/domain/makeup/repository/MakeUpRepository.java
@@ -1,7 +1,7 @@
-package spring.beautiq.makeUp.repository;
+package spring.beautiq.domain.makeup.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spring.beautiq.makeUp.entity.MakeUp;
+import spring.beautiq.domain.makeup.entity.MakeUp;
 
 public interface MakeUpRepository extends JpaRepository<MakeUp, Long> {
 }
